Fix babel runtime resolution issue
