+++ extensions/cornerstone/package.json
@@ -43,7 +43,7 @@
     "redux": "^4.0.1"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0",
+    "@babel/runtime": "^7.5.5",
     "classnames": "^2.2.6",
     "lodash.throttle": "^4.1.1",
     "react-cornerstone-viewport": "0.1.30"
