+++ extensions/dicom-html/package.json
@@ -33,6 +33,6 @@
     "react-dom": "^15.0.0 || ^16.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0"
+    "@babel/runtime": "^7.5.5"
   }
 }
