+++ extensions/dicom-microscopy/package.json
@@ -31,7 +31,7 @@
     "react-dom": "^15.0.0 || ^16.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0",
+    "@babel/runtime": "^7.5.5",
     "classnames": "^2.2.6",
     "dicom-microscopy-viewer": "^0.4.3"
   }
