+++ extensions/dicom-pdf/package.json
@@ -33,7 +33,7 @@
     "react-dom": "^15.0.0 || ^16.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0",
+    "@babel/runtime": "^7.5.5",
     "classnames": "^2.2.6",
     "lodash.isequal": "^4.5.0"
   },
