+++ platform/core/package.json
@@ -11,20 +11,23 @@
   "publishConfig": {
     "access": "public"
   },
+  "files": [
+    "dist",
+    "README.md"
+  ],
   "engines": {
     "node": ">=8",
     "npm": ">=5"
   },
   "scripts": {
-    "start": "yarn run dev",
-    "dev": "jest --watch",
+    "dev": "jest --watchAll",
     "dev:core": "yarn run dev",
-    "test:unit": "jest --watch",
-    "test:unit:ci": "jest --ci --runInBand --collectCoverage",
-    "build": "webpack --config config/webpack.prod.js -d",
+    "build": "webpack --config .webpack/webpack.prod.js",
     "build:package": "yarn run build",
-    "lint": "eslint -c .eslintrc --fix src && prettier --single-quote --write src/**/*.{js,jsx,json,css}",
-    "prepublishOnly": "yarn run build"
+    "prepublishOnly": "yarn run build",
+    "start": "yarn run dev",
+    "test:unit": "jest --watchAll",
+    "test:unit:ci": "jest --ci --runInBand --collectCoverage"
   },
   "peerDependencies": {
     "cornerstone-core": "^2.2.8",
@@ -33,7 +36,7 @@
     "dicom-parser": "^1.8.3"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0",
+    "@babel/runtime": "^7.5.5",
     "ajv": "^6.10.0",
     "dicomweb-client": "^0.4.2",
     "isomorphic-base64": "^1.0.2",
@@ -41,10 +44,5 @@
     "lodash.merge": "^4.6.1",
     "mousetrap": "^1.6.3",
     "validate.js": "^0.12.0"
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "gitHead": "a5baa9228c0eda0df880136bde4420d78e6f8706"
+  }
 }
