+++ platform/i18n/package.json
@@ -19,14 +19,14 @@
     "access": "public"
   },
   "scripts": {
-    "test:unit": "echo 'platform/i18n: missing unit tests'",
-    "test:unit:ci": "echo 'platform/i18n: missing unit tests'",
-    "dev": "webpack --config config/webpack.prod.js -w -d",
+    "dev": "webpack --config .webpack/webpack.dev.js --watch --debug --output-pathinfo",
     "dev:i18n": "yarn run dev",
-    "prepublishOnly": "yarn run build",
-    "build": "webpack --config config/webpack.prod.js -d",
+    "build": "webpack --config .webpack/webpack.prod.js",
     "build:package": "yarn run build",
-    "pullTranslations": "./pullTranslations.sh"
+    "prepublishOnly": "yarn run build",
+    "pullTranslations": "./pullTranslations.sh",
+    "test:unit": "echo 'platform/i18n: missing unit tests'",
+    "test:unit:ci": "echo 'platform/i18n: missing unit tests'"
   },
   "peerDependencies": {
     "i18next": "^17.0.3",
@@ -36,7 +36,7 @@
     "react-i18next": "^10.11.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.2.0",
+    "@babel/runtime": "^7.5.5",
     "classnames": "^2.2.6",
     "i18next-locize-backend": "^2.0.0",
     "locize-editor": "^2.0.0",
@@ -48,6 +48,5 @@
     "i18next-browser-languagedetector": "^3.0.1",
     "locize-cli": "^4.8.0",
     "react-i18next": "^10.11.0"
-  },
-  "gitHead": "a5baa9228c0eda0df880136bde4420d78e6f8706"
+  }
 }
